refactors sign up validations as middlewares and renames userServices to userMiddleware
